+++ test/clj_book/compose_test.clj
@@ -1,20 +1,13 @@
 (ns clj-book.compose-test
   (:require
    [clj-book.compose :as compose]
-   [clj-book.error :as error]
    [clojure.java.io :as io]
    [clojure.string :as str]
    [clojure.test :refer [deftest is testing]]))
 
-(def valid-root "test/fixtures/synthetic/valid-book")
-
-(defn- catch-data [f]
-  (try (f) nil (catch Exception e (error/data e))))
-
 (deftest master-adoc-respects-chapter-order
   (let [s (compose/master-adoc
-            {:book-root valid-root
-             :config {:book/slug "x"
+            {:config {:book/slug "x"
                       :book/title "X"
                       :book/chapters ["chapters/02-body.adoc"
                                       "chapters/01-intro.adoc"]}})
@@ -30,33 +23,13 @@
                :book/title "X"
                :book/chapters ["chapters/01-intro.adoc"
                                "chapters/02-body.adoc"]}
-          a (compose/master-adoc {:book-root valid-root :config cfg})
-          b (compose/master-adoc {:book-root valid-root :config cfg})]
+          a (compose/master-adoc {:config cfg})
+          b (compose/master-adoc {:config cfg})]
       (is (= a b)))))
-
-(deftest duplicate-chapters-fail
-  (let [d (catch-data
-            #(compose/master-adoc
-               {:book-root valid-root
-                :config {:book/slug "x"
-                         :book/title "X"
-                         :book/chapters ["chapters/01-intro.adoc"
-                                         "chapters/01-intro.adoc"]}}))]
-    (is (= :clj-book.compose/duplicate-chapter (:error/type d)))))
-
-(deftest missing-chapter-fails
-  (let [d (catch-data
-            #(compose/master-adoc
-               {:book-root valid-root
-                :config {:book/slug "x"
-                         :book/title "X"
-                         :book/chapters ["chapters/missing.adoc"]}}))]
-    (is (= :clj-book.compose/missing-chapter (:error/type d)))))
 
 (deftest master-includes-header-title
   (let [s (compose/master-adoc
-            {:book-root valid-root
-             :config {:book/slug "x"
+            {:config {:book/slug "x"
                       :book/title "Great Book"
                       :book/chapters ["chapters/01-intro.adoc"]}})]
     (is (str/starts-with? s "= Great Book"))
@@ -66,8 +39,7 @@
   (let [tmp (str (System/getProperty "java.io.tmpdir")
                  "/clj-book-compose-test-" (System/currentTimeMillis))
         path (compose/write-master!
-               {:book-root valid-root
-                :intermediate-dir tmp
+               {:intermediate-dir tmp
                 :config {:book/slug "x"
                          :book/title "X"
                          :book/chapters ["chapters/01-intro.adoc"]}})]